kafka: move flex version bits to kafka protocol
These are used both by the server and client.

ld.lld: error: undefined symbol: kafka::flex_versions::is_api_in_schema(detail::base_named_...
>>> referenced by transport.h:62 (../src/v/kafka/client/transport.h:62)
>>>               src/v/kafka/client/libkafka_client.so.p/client.cc.o:(auto kafka::client::...

ld.lld: error: undefined symbol: kafka::parse_tags(seastar::input_stream<char>&)
>>> referenced by transport.h:94 (../src/v/kafka/client/transport.h:94)
>>>               src/v/kafka/client/libkafka_client.so.p/client.cc.o:(auto kafk...
