+++ src/v/kafka/CMakeLists.txt
@@ -55,7 +55,6 @@ v_cc_library(
     server/connection_context.cc
     server/protocol.cc
     server/protocol_utils.cc
-    server/flex_versions.cc
     server/quota_manager.cc
     server/fetch_session_cache.cc
     server/replicated_partition.cc
