+++ src/v/kafka/server/protocol_utils.cc
@@ -11,7 +11,7 @@
 
 #include "bytes/iobuf.h"
 #include "bytes/iobuf_parser.h"
-#include "kafka/server/flex_versions.h"
+#include "kafka/protocol/flex_versions.h"
 
 #include <seastar/core/temporary_buffer.hh>
 
@@ -84,44 +84,6 @@ parse_header(ss::input_stream<char>& src) {
         header.tags_size_bytes = bytes_read;
     }
     co_return header;
-}
-
-ss::future<std::pair<std::optional<tagged_fields>, size_t>>
-parse_tags(ss::input_stream<char>& src) {
-    size_t total_bytes_read = 0;
-    auto read_unsigned_vint =
-      [](size_t& total_bytes_read, ss::input_stream<char>& src) {
-          return unsigned_vint::stream_deserialize(src).then(
-            [&total_bytes_read](std::pair<uint32_t, size_t> pair) {
-                auto& [n, bytes_read] = pair;
-                total_bytes_read += bytes_read;
-                return n;
-            });
-      };
-
-    auto num_tags = co_await read_unsigned_vint(total_bytes_read, src);
-    if (num_tags == 0) {
-        /// In the likely event that no tags are parsed as headers, return
-        /// nullopt instead of an empty vector to reduce the memory overhead
-        /// (that will never be used) per request
-        co_return std::make_pair(std::nullopt, total_bytes_read);
-    }
-
-    tagged_fields::type tags;
-    while (num_tags-- > 0) {
-        auto id = co_await read_unsigned_vint(total_bytes_read, src);
-        auto next_len = co_await read_unsigned_vint(total_bytes_read, src);
-        auto buf = co_await src.read_exactly(next_len);
-        auto data = ss::uninitialized_string<bytes>(buf.size());
-        std::copy_n(buf.begin(), buf.size(), data.begin());
-        total_bytes_read += next_len;
-        auto [_, succeded] = tags.emplace(tag_id(id), std::move(data));
-        if (!succeded) {
-            throw std::logic_error(
-              fmt::format("Protocol error, duplicate tag id detected, {}", id));
-        }
-    }
-    co_return std::make_pair(std::move(tags), total_bytes_read);
 }
 
 size_t parse_size_buffer(ss::temporary_buffer<char> buf) {
