+++ src/v/kafka/server/protocol_utils.h
@@ -24,9 +24,6 @@
 
 namespace kafka {
 
-ss::future<std::pair<std::optional<tagged_fields>, size_t>>
-parse_tags(ss::input_stream<char>&);
-
 // TODO: move to iobuf_parser
 ss::future<std::optional<request_header>> parse_header(ss::input_stream<char>&);
